Create canWinNim.java
Problem number: 292. Nim Game

The code snippet contains two alternative implementations of the "Nim Game" problem in Java. The Nim Game is a classic combinatorial game where two players take turns removing stones from a pile, and the goal is to avoid being the player to take the last stone(s). These implementations determine whether a player can win the Nim Game based on the number of stones in the initial pile.