Create climbStairs.java
The given code is a Java implementation of a solution to the classic "Climbing Stairs" problem. This problem involves finding the number of distinct ways a person can climb a staircase with n steps. The person can climb either 1 or 2 steps at a time. The code employs an iterative approach to calculate this count efficiently.
* Another way is to use Dynamic Programming. That will be updated shortly.